pyproject.toml
[project]
name = "simple-deep-research"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepagents>=0.0.3",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.30",
    "langgraph-cli[inmem]>=0.3.6",
    "python-dotenv>=1.1.1",
    "tavily-python>=0.7.11",
]
